Always export GPG key to deb mirror folder in update_mirror_deb.sh
If the key is not exported and the key in the folder came from another machine, i.e. the mirror files were generated elsewhere, then the key used to sign the packages as part of the update operation will not match the one available for clients to download. For this reason, always export the key used when updating the mirror.

PNDA-3078